+++ src/test/java/mil/nga/geopackage/ReadmeTest.java
@@ -2,6 +2,8 @@ package mil.nga.geopackage;
 
 import java.awt.image.BufferedImage;
 import java.io.File;
+import java.io.IOException;
+import java.sql.SQLException;
 import java.util.List;
 
 import org.junit.Test;
@@ -11,11 +13,12 @@ import mil.nga.geopackage.extension.ExtensionsDao;
 import mil.nga.geopackage.extension.metadata.MetadataDao;
 import mil.nga.geopackage.extension.metadata.MetadataExtension;
 import mil.nga.geopackage.extension.metadata.reference.MetadataReferenceDao;
-import mil.nga.geopackage.extension.nga.index.FeatureTableIndex;
 import mil.nga.geopackage.extension.schema.SchemaExtension;
 import mil.nga.geopackage.extension.schema.columns.DataColumnsDao;
 import mil.nga.geopackage.extension.schema.constraints.DataColumnConstraintsDao;
 import mil.nga.geopackage.features.columns.GeometryColumnsDao;
+import mil.nga.geopackage.features.index.FeatureIndexManager;
+import mil.nga.geopackage.features.index.FeatureIndexType;
 import mil.nga.geopackage.features.user.FeatureDao;
 import mil.nga.geopackage.features.user.FeatureResultSet;
 import mil.nga.geopackage.features.user.FeatureRow;
@@ -25,11 +28,18 @@ import mil.nga.geopackage.tiles.GeoPackageTile;
 import mil.nga.geopackage.tiles.GeoPackageTileRetriever;
 import mil.nga.geopackage.tiles.ImageUtils;
 import mil.nga.geopackage.tiles.TileCreator;
+import mil.nga.geopackage.tiles.TileGenerator;
+import mil.nga.geopackage.tiles.UrlTileGenerator;
+import mil.nga.geopackage.tiles.features.DefaultFeatureTiles;
+import mil.nga.geopackage.tiles.features.FeatureTileGenerator;
+import mil.nga.geopackage.tiles.features.FeatureTiles;
+import mil.nga.geopackage.tiles.features.custom.NumberFeaturesTile;
 import mil.nga.geopackage.tiles.matrix.TileMatrixDao;
 import mil.nga.geopackage.tiles.matrixset.TileMatrixSetDao;
 import mil.nga.geopackage.tiles.user.TileDao;
 import mil.nga.geopackage.tiles.user.TileResultSet;
 import mil.nga.geopackage.tiles.user.TileRow;
+import mil.nga.proj.Projection;
 import mil.nga.proj.ProjectionConstants;
 import mil.nga.proj.ProjectionFactory;
 import mil.nga.sf.Geometry;
@@ -43,9 +53,14 @@ public class ReadmeTest extends CreateGeoPackageTestCase {
 
 	/**
 	 * Test transform
+	 * 
+	 * @throws IOException
+	 *             upon error
+	 * @throws SQLException
+	 *             upon error
 	 */
 	@Test
-	public void testGeoPackage() {
+	public void testGeoPackage() throws SQLException, IOException {
 
 		File newGeoPackage = new File("new.gpkg");
 		newGeoPackage.delete();
@@ -68,8 +83,13 @@ public class ReadmeTest extends CreateGeoPackageTestCase {
 	 *            new GeoPackage
 	 * @param existingGeoPackage
 	 *            existing GeoPackage
+	 * @throws IOException
+	 *             upon error
+	 * @throws SQLException
+	 *             upon error
 	 */
-	private void testGeoPackage(File newGeoPackage, File existingGeoPackage) {
+	private void testGeoPackage(File newGeoPackage, File existingGeoPackage)
+			throws SQLException, IOException {
 
 		// File newGeoPackage = ...;
 		// File existingGeoPackage = ...;
@@ -102,7 +122,8 @@ public class ReadmeTest extends CreateGeoPackageTestCase {
 		List<String> tiles = geoPackage.getTileTables();
 
 		// Query Features
-		FeatureDao featureDao = geoPackage.getFeatureDao(features.get(0));
+		String featureTable = features.get(0);
+		FeatureDao featureDao = geoPackage.getFeatureDao(featureTable);
 		FeatureResultSet featureResultSet = featureDao.queryForAll();
 		try {
 			while (featureResultSet.moveToNext()) {
@@ -118,12 +139,14 @@ public class ReadmeTest extends CreateGeoPackageTestCase {
 		}
 
 		// Query Tiles
-		TileDao tileDao = geoPackage.getTileDao(tiles.get(0));
+		String tileTable = tiles.get(0);
+		TileDao tileDao = geoPackage.getTileDao(tileTable);
 		TileResultSet tileResultSet = tileDao.queryForAll();
 		try {
 			while (tileResultSet.moveToNext()) {
 				TileRow tileRow = tileResultSet.getRow();
 				byte[] tileBytes = tileRow.getTileData();
+				BufferedImage tileImage = tileRow.getTileDataImage();
 				// ...
 			}
 		} finally {
@@ -135,6 +158,7 @@ public class ReadmeTest extends CreateGeoPackageTestCase {
 				ImageUtils.IMAGE_FORMAT_PNG);
 		GeoPackageTile geoPackageTile = retriever.getTile(2, 2, 2);
 		if (geoPackageTile != null) {
+			byte[] tileBytes = geoPackageTile.getData();
 			BufferedImage tileImage = geoPackageTile.getImage();
 			// ...
 		}
@@ -147,14 +171,47 @@ public class ReadmeTest extends CreateGeoPackageTestCase {
 		GeoPackageTile geoPackageTile2 = tileCreator
 				.getTile(new BoundingBox(-90.0, 0.0, 0.0, 66.513260));
 		if (geoPackageTile2 != null) {
+			byte[] tileBytes = geoPackageTile2.getData();
 			BufferedImage tileImage = geoPackageTile2.getImage();
 			// ...
 		}
 
 		// Index Features
-		FeatureTableIndex indexer = new FeatureTableIndex(geoPackage,
+		FeatureIndexManager indexer = new FeatureIndexManager(geoPackage,
 				featureDao);
+		indexer.setIndexLocation(FeatureIndexType.GEOPACKAGE);
 		int indexedCount = indexer.index();
+
+		// Draw tiles from features
+		FeatureTiles featureTiles = new DefaultFeatureTiles(featureDao);
+		// Set max features to draw per tile
+		featureTiles.setMaxFeaturesPerTile(1000);
+		// Custom feature tile implementation
+		NumberFeaturesTile numberFeaturesTile = new NumberFeaturesTile();
+		// Draw feature count tiles when max features passed
+		featureTiles.setMaxFeaturesTileDraw(numberFeaturesTile);
+		// Set index manager to query feature indices
+		featureTiles.setIndexManager(indexer);
+		BufferedImage tile = featureTiles.drawTile(2, 2, 2);
+
+		BoundingBox boundingBox = BoundingBox.worldWebMercator();
+		Projection projection = ProjectionFactory
+				.getProjection(ProjectionConstants.EPSG_WEB_MERCATOR);
+
+		// URL Tile Generator (generate tiles from a URL)
+		TileGenerator urlTileGenerator = new UrlTileGenerator(geoPackage,
+				"url_tile_table", "http://url/{z}/{x}/{y}.png", 0, 0,
+				boundingBox, projection);
+		int urlTileCount = urlTileGenerator.generateTiles();
+
+		// Feature Tile Generator (generate tiles from features)
+		TileGenerator featureTileGenerator = new FeatureTileGenerator(
+				geoPackage, "tiles_" + featureTable, featureTiles, 1, 2,
+				boundingBox, projection);
+		int featureTileCount = featureTileGenerator.generateTiles();
+
+		// Close feature tiles (and indexer)
+		featureTiles.close();
 
 		// Close database when done
 		geoPackage.close();